OF-2569 FAST - HRefactor HT-/HT2- to common base
